+++ src/settings.ts
@@ -4,30 +4,46 @@ export interface ISettings
     IndentLength: number;
     UseSingleQuotes: boolean;
     UseSelfClosingTags: boolean;
-    FormatOnSave : boolean;
+    FormatOnSave: boolean;
+    AllowSingleQuoteInAttributeValue: boolean;
+    AddSpaceBeforeSelfClosingTag: boolean;
+    WrapCommentTextWithSpaces: boolean;
 }
 
 export const DefaultSettings: ISettings = {
     IndentLength: 2,
     UseSingleQuotes: false,
     UseSelfClosingTags: true,
-    FormatOnSave : false,
+    FormatOnSave: false,
+    AllowSingleQuoteInAttributeValue: true,
+    AddSpaceBeforeSelfClosingTag: true,
+    WrapCommentTextWithSpaces: true
 };
 
 export class Settings
 {
     IndentLength?: number;
     UseSingleQuotes?: boolean;
-    UseSelfClosingTags?: boolean ;
-    FormatOnSave?:boolean;
+    UseSelfClosingTags?: boolean;
+    FormatOnSave?: boolean;
+    AllowSingleQuoteInAttributeValue?: boolean;
+    AddSpaceBeforeSelfClosingTag?: boolean;
+    WrapCommentTextWithSpaces?: boolean;
 
-    constructor(indentLengh?:number, useSingleQuotes?:boolean, useSelfClosingTags?:boolean, formatOnSave?:boolean)
+    constructor(indentLengh?: number,
+        useSingleQuotes?: boolean,
+        useSelfClosingTags?: boolean,
+        formatOnSave?: boolean,
+        allowSingleQuoteInAttributeValue?: boolean,
+        addSpaceBeforeSelfClosingTag?: boolean,
+        wrapCommentTextWithSpaces?: boolean)
     {
         this.IndentLength = indentLengh ?? DefaultSettings.IndentLength;
         this.UseSingleQuotes = useSingleQuotes ?? DefaultSettings.UseSingleQuotes;
-        this.UseSelfClosingTags = useSelfClosingTags?? DefaultSettings.UseSelfClosingTags;
+        this.UseSelfClosingTags = useSelfClosingTags ?? DefaultSettings.UseSelfClosingTags;
         this.FormatOnSave = formatOnSave ?? DefaultSettings.FormatOnSave;
+        this.AllowSingleQuoteInAttributeValue = allowSingleQuoteInAttributeValue;
+        this.AddSpaceBeforeSelfClosingTag = addSpaceBeforeSelfClosingTag ?? DefaultSettings.AddSpaceBeforeSelfClosingTag;
+        this.WrapCommentTextWithSpaces = wrapCommentTextWithSpaces ?? DefaultSettings.WrapCommentTextWithSpaces;
     }
 }
-
-